local-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.0
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
pywin32==306
watchdog==3.0.0
